added getScheduledJobs, second try
Sorry @pitchmuc , I probably should have done sth else instead of closing and then reopening another Pull Request... This should now work, the last draft had a misunderstanding in it in terms of how optional parameters work when you use the "hint" notation that I am not too familiar with yet.

The text is still valid: 
This allows you to get all scheduled projects. Can be useful to monitor expired or not delivered schedules or to combine with workspaces (soon to come) to see which workspaces have schedules attached to them.