Update token_provider.py
Right now, the handling of failed logins is suboptimal: The script prints the error, but continues with evaluating the json_response, which then fails because there is obviously no `expires_in` property in this case. 
The function should return the erroneous json_response or raise an Exception so it can be propagated to the calling function (and eg shown to the user of the application). 
Right now the users get a very unhelpful error message, it is only:"'expires_in'". With a proper Exception, the error for eg false jwt credentials is: "{'error_description': 'Could not match JWT signature to any of the bindings', 'error': 'invalid_token'}" => much more helpful