Allow getProjects with custom expansion parameters
Right now, the "expansion" parameter for the `getProjects` method is always either the `full` list of `'reportSuiteName,ownerFullName,tags,shares,sharesFullName,modified,favorite,approved,companyTemplate,externalReferences,accessLevel'` or the smaller subset of: 

```            
            params["expansion"] = "ownerFullName,modified"
            if includeShared:
                params["expansion"] += ',shares,sharesFullName'
            if includeTemplate:
                params["expansion"] += ',companyTemplate'
```

In some cases where clients have 10s of thousands of projects, you can speed up the query by limiting the expansion parameters.

Also, the "externalReferences" expansion parameter causes the getProjects request to fail for a client of mine ("Connection broken"). Since `externalReferences` does not contain anything of value for us, we could live without it, but for that we need a way to query getProjects with a custom expansion parameter list.

This simple change will allow that. So thanks for looking at this, @pitchmuc