Avoid keyError: "retry" in case request fails
If a request fails completely (eg "Connection broken" due to giant segments), the connector.py tries again. However, the second time this happens, there is a conflict between the retry parameter in kwargs and the separate retry parameter which leads to a "keyError: 'retry'" error.
This can be avoided via the changes here. I also set time.sleep to the  connector.restTime so this is consistent with the rest of the retries.